+++ plone/app/linkintegrity/browser/update.py
@@ -5,30 +5,55 @@ from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import PloneMessageFactory as _
 from Products.Five import BrowserView
 from Products.statusmessages.interfaces import IStatusMessage
-
+from datetime import datetime
+from datetime import timedelta
 from plone.app.linkintegrity.handlers import modifiedArchetype
 from plone.app.linkintegrity.handlers import modifiedDexterity
-from plone.app.linkintegrity import HAS_LINGUAPLONE
-from plone.app.linkintegrity import HAS_PAM
-from plone.app.referenceablebehavior.referenceable import IReferenceable
 from plone.dexterity.interfaces import IDexterityContent
+from zExceptions import NotFound
+import logging
+import pkg_resources
+
+# Is there a multilingual addon?
+try:
+    pkg_resources.get_distribution('Products.LinguaPlone')
+except pkg_resources.DistributionNotFound:
+    HAS_MULTILINGUAL = False
+else:
+    HAS_MULTILINGUAL = True
+
+if not HAS_MULTILINGUAL:
+    try:
+        pkg_resources.get_distribution('plone.app.multilingual')
+    except pkg_resources.DistributionNotFound:
+        HAS_MULTILINGUAL = False
+    else:
+        HAS_MULTILINGUAL = True
+
+logger = logging.getLogger(__name__)
 
 
 class UpdateView(BrowserView):
+    """Iterate over all catalogued items and update linkintegrity-information.
+    """
 
     def __call__(self):
         context = aq_inner(self.context)
         request = aq_inner(self.request)
         clicked = request.form.has_key
         if clicked('update') or clicked('delete_all'):
+            starttime = datetime.now()
             count = self.update()
+            duration = timedelta(seconds=(datetime.now() - starttime).seconds)
             msg = _(
                 u'linkintegrity_update_info',
                 default=u'Link integrity information updated for ${count} ' +
-                        u'item(s).',
-                mapping={'count': count},
+                        u'items in ${time} seconds.',
+                mapping={'count': count, 'time': str(duration)},
             )
             IStatusMessage(request).add(msg, type='info')
+            msg = 'Updated {} items in {} seconds'.format(count, str(duration))
+            logger.info(msg)
             request.RESPONSE.redirect(getToolByName(context, 'portal_url')())
         elif clicked('cancel'):
             msg = _(u'Update cancelled.')
@@ -40,18 +65,27 @@ class UpdateView(BrowserView):
     def update(self):
         catalog = getToolByName(self.context, 'portal_catalog')
         count = 0
-        kwargs = {}
+        query = {}
+        if HAS_MULTILINGUAL and 'Language' in catalog.indexes():
+            query['Language'] = 'all'
 
-        if HAS_LINGUAPLONE or HAS_PAM:
-            kwargs['Language'] = 'all'
-
-        for brain in catalog(**kwargs):
-            obj = brain.getObject()
+        for brain in catalog(query):
+            try:
+                obj = brain.getObject()
+            except (AttributeError, NotFound, KeyError):
+                msg = "Catalog inconsistency: {} not found!"
+                logger.error(msg.format(brain.getPath()), exc_info=1)
+                continue
+            method = None
             if IBaseObject.providedBy(obj):
-                modifiedArchetype(obj, 'dummy event parameter')
-                count += 1
+                method = modifiedArchetype
             elif IDexterityContent.providedBy(obj):
-                if IReferenceable.providedBy(obj):
-                    modifiedDexterity(obj, 'dummy event parameter')
-                count += 1
+                method = modifiedDexterity
+            if method:
+                try:
+                    method(obj, 'dummy event parameter')
+                    count += 1
+                except Exception:
+                    msg = "Error updating linkintegrity-info for {}."
+                    logger.error(msg.format(obj.absolute_url()), exc_info=1)
         return count