+++ plone/app/linkintegrity/utils.py
@@ -1,52 +1,65 @@
-from zExceptions import BadRequest
-from base64 import b64encode, b64decode
-from zlib import compress, decompressobj
+# -*- coding: utf-8 -*-
+from Products.CMFPlone.interfaces import IEditingSchema
+from plone.app.linkintegrity.handlers import referencedRelationship
+from plone.registry.interfaces import IRegistry
+from zc.relation.interfaces import ICatalog
+from zope.component import getUtility
+from zope.intid.interfaces import IIntIds
 
 
-def decompress(data, maxsize=262144):
-
-    dec = decompressobj()
-    data = dec.decompress(data, maxsize)
-    if dec.unconsumed_tail:
-        raise BadRequest
-    del dec
-
-    return data
-
-
-def encodeStrings(strings):
-    """ compress and encode a list of strings into a single string """
-    def _encode(strings):
-        for string in strings:
-            yield '%d:%s' % (len(string), string)
-    return b64encode(compress(''.join(_encode(strings))))
+def getIncomingLinks(obj=None, intid=None):
+    """Return a generator of incoming relations created using
+    plone.app.linkintegrity (Links in Richtext-Fields).
+    """
+    catalog = getUtility(ICatalog)
+    if intid is not None:
+        return catalog.findRelations({
+            'to_id': intid,
+            'from_attribute': referencedRelationship})
+    else:
+        intids = getUtility(IIntIds)
+        return catalog.findRelations({
+            'to_id': intids.getId(obj),
+            'from_attribute': referencedRelationship})
 
 
-def decodeStrings(data):
-    """ decode and uncompress a string as a generator """
-    data = decompress(b64decode(data))
-    while data:
-        pos = data.find(':')
-        end = pos + int(data[:pos]) + 1
-        yield data[pos + 1:end]
-        data = data[end:]
+def hasIncomingLinks(obj=None, intid=None):
+    """Test if an object is linked to by other objects using
+    plone.app.linkintegrity (Links in Richtext-Fields).
+
+    Way to give bool without loading generator into list
+    """
+    for i in getIncomingLinks(obj=obj, intid=intid):
+        return True
+    return False
 
 
-def encodeRequestData((body, env)):
-    """ encode the relevant request data, i.e. body and env """
-    def _iterdata():
-        yield body
-        for key, val in env.iteritems():
-            if isinstance(key, basestring) and isinstance(val, basestring):
-                yield str(key)
-                yield str(val)
-    return encodeStrings(_iterdata())
+def getOutgoingLinks(obj=None, intid=None):
+    """Return a generator of outgoing relations created using
+    plone.app.linkintegrity (Links in Richtext-Fields).
+    """
+    catalog = getUtility(ICatalog)
+    if intid is not None:
+        return catalog.findRelations({
+            'from_id': intid,
+            'from_attribute': referencedRelationship})
+    else:
+        intids = getUtility(IIntIds)
+        return catalog.findRelations({
+            'from_id': intids.getId(obj),
+            'from_attribute': referencedRelationship})
 
 
-def decodeRequestData(data):
-    """ decode the relevant request data, i.e. body and env """
-    def _pertwo():
-        while True:
-            yield data.next(), data.next()
-    data = decodeStrings(data)
-    return data.next(), dict(_pertwo())
+def hasOutgoingLinks(obj=None, intid=None):
+    """Test if an object links to other objects using plone.app.linkintegrity
+    (Links in Richtext-Fields).
+    """
+    for i in getOutgoingLinks(obj=obj, intid=intid):
+        return True
+    return False
+
+
+def linkintegrity_enabled():
+    reg = getUtility(IRegistry)
+    editing_settings = reg.forInterface(IEditingSchema, prefix='plone')
+    return editing_settings.enable_link_integrity_checks